- Input security check before deleting mmap files

+++ custom scripts/div/file_manager.py
@@ -48,6 +48,8 @@ def remove_mmap_after_analysis(root):
     """
     free_mem = 0
     n_files = 0
+
+    del_files = []
     for step in os.walk(root):
         mmap_file = glob(step[0]+'\\memmap__*')
         ana_file = glob(step[0] + '\\pcf_result*')
@@ -55,5 +57,19 @@ def remove_mmap_after_analysis(root):
             for file in mmap_file:
                 free_mem += os.path.getsize(file)/1073741824    # Remember size of deleted file for later report
                 n_files += 1
+                del_files.append(file)
+
+    print(f'Found {len(del_files)} files to be deleted:')
+    print(*del_files, sep='\n')
+    answer = None
+    while answer not in ("y", "n", 'yes', 'no'):
+        answer = input(f'These files would free up {int(free_mem*100)/100} GB. Do you want to delete them? [y/n]')
+        if answer == "yes" or answer == 'y':
+            print('Deleting...')
+            for file in del_files:
                 os.remove(file)
-    print(f'Done! Deleting {n_files} mmap files freed up ca. {int(free_mem*100)/100} GB on {root}.')
+            print('Done!')
+        elif answer == "no" or answer == 'n':
+            print('Deleting cancelled.')
+        else:
+            print("Please enter yes or no.")
